pyproject.toml
[tool.poetry]
name = "kinocms"
version = "0.1.0"
description = ""
authors = ["dhushchyn <dhushchyn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"
django-meta = "^2.4.2"
gunicorn = "^21.2.0"
django-cors-headers = "^4.3.1"
pydantic = {extras = ["email"], version = "^2.6.3"}
django-ninja-jwt = "^5.3.0"
celery = "^5.3.6"
redis = "^5.0.2"
django-cleanup = "^8.1.0"
pre-commit = "^3.6.2"
django-modeltranslation = "^0.18.11"
openapi = "^1.1.0"
django-imagekit = "^5.0.0"
faker = "^24.4.0"
django-extensions = "^3.2.3"
pytest = "^8.1.1"
pytest-django = "^4.8.0"
loguru = "^0.7.2"
django-redis = "^5.4.0"
pendulum = "^3.0.0"
translate = "^3.6.1"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.3.0"}
pydantic-extra-types = "^2.7.0"
pyjwt = "^2.8.0"
django-ninja-extra = "^0.20.7"
ninja-schema = "^0.13.6"
pytils = "^0.4.1"
django-countries = "^7.6.1"
requests-tracker = "^0.3.3"
django-multiselectfield = "^0.1.12"
matplotlib = "^3.9.0"
pytz = "^2024.1"
fastapi-errors = "^0.0.6"
babel = "^2.15.0"
pymorphy2 = "^0.9.1"
pymorphy2-dicts-uk = "^2.4.1.1.1460299261"
requests-toolbelt = "^1.0.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



# ==== Coverage ====
[tool.coverage.run]
include = ["my_awesome_project/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "pypackages",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "*/migrations/*.py",
    "staticfiles/*"
]
# Same as Django: https://github.com/cookiecutter/cookiecutter-django/issues/4792.
line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "C90",
  "I",
  "N",
  "UP",
  "YTT",
  # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
  "ASYNC",
  "D",
  "S",
  "BLE",
  "FBT",
  "B",
  "A",
  "COM",
  "C4",
  "DTZ",
  "T10",
  "DJ",
  "EM",
  "EXE",
  "FA",
  'ISC',
  "ICN",
  "G",
  'INP',
  'PIE',
  "T20",
  'PYI',
  'PT',
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "SIM",
  "TID",
  "TCH",
  "INT",
  # "ARG", # Unused function argument
  "PTH",
  "ERA",
  "PD",
  "PGH",
  "PL",
  "TRY",
  "FLY",
  # "NPY",
  # "AIR",
  "PERF",
  # "FURB",
  # "LOG",
  "RUF"
]
ignore = [
  "S101", # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "RUF012", # Mutable class attributes should be annotated with typing.ClassVar
  "SIM102", # sometimes it's better to nest
  "UP038", # Checks for uses of isinstance/issubclass that take a tuple
          # of types for comparison.
          # Deactivated because it can make the code slow:
          # https://github.com/astral-sh/ruff/issues/7871
    "D203", #  one-blank-line-before-class
    "D213", #  `multi-line-summary-second-line
    "COM812",
    "ISC001"
]
# Allow fix for all enabled rules (when --fix) is provided.
fixable = ["ALL"]
unfixable = []
# The fixes in extend-unsafe-fixes will require
# provide the --unsafe-fixes flag when fixing.
extend-unsafe-fixes = [
    "UP038"
]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
force-single-line = true
